Move costfunction and manifold in dedicated files
