+++ src/aliceVision/sfm/BundleAdjustmentSymbolicCeres.hpp
@@ -9,13 +9,15 @@
 
 #include <aliceVision/types.hpp>
 #include <aliceVision/alicevision_omp.hpp>
+#include <aliceVision/numeric/numeric.hpp>
+
 #include <aliceVision/sfm/BundleAdjustment.hpp>
 #include <aliceVision/sfm/LocalBundleAdjustmentGraph.hpp>
-#include <aliceVision/numeric/numeric.hpp>
+
 #include <aliceVision/sfmData/CameraPose.hpp>
 #include <aliceVision/camera/IntrinsicBase.hpp>
+
 #include <ceres/ceres.h>
-#include "liealgebra.hpp"
 
 #include <memory>
